When I merged the massive amount of work in #6289, I thought that that
was the end of it. But simply staging and unpacking loose blobs that are
4GB or larger _still_ does not work. This PR addresses that.

This fixes https://github.com/git-for-windows/git/issues/6012

+++ odb.c
@@ -980,7 +980,7 @@ void odb_assert_oid_type(struct object_database *odb,
 }
 
 int odb_write_object_ext(struct object_database *odb,
-			 const void *buf, unsigned long len,
+			 const void *buf, size_t len,
 			 enum object_type type,
 			 struct object_id *oid,
 			 const struct object_id *compat_oid_in,
